sbull.c: Change elv_next_request to blk_peek_request
Following a LWN article[1], elv_next_request is past, and now drivers
should use blk_peek_request instead. The former is known to return a
request from the queue, but not removing it. The later return the
request, and removes it from the queue, leaving the driver to take care
of it.

Also, elv_next_request is not available anymore to drives to be used.

[1]: https://lwn.net/Articles/333620/

Signed-off-by: Marcos Paulo de Souza <marcos.souza.org@gmail.com>