sbull-mq: Rework sector counter and fix capacity of sbull
Until now sbull was used the "nsectors" approach to define the size of
disk, and this somehow led to a wrong assumption of sbull capacity.

set_capacity's argument is the number of sectors of "512 bytes" a driver
has, so, for a 32 MB of size, the following approach is needed:

32 MB == 33554432 bytes
33554432 bytes / 512 == 65535 blocks of "512 bytes"

Until now, the nsectors module parameter was using 65535 sectors, which
always led to a 32 MB of RAM. It wasn't a bug for itself, but rather a
concept not understood by me.

This patch makes the following changes:

* The module argument called nsectors is removed, and disk_size is
added. This new argument can receive a string, like "256M", which will
make sbull to have a disk with 256M of space.

* Two new macros are added to calculate the "number of bytes / number of
sectors", and the "sector -> byte"

Now sbull can work with other filesystems, like ext4 and xfs.

Signed-off-by: Marcos Paulo de Souza <marcos.souza.org@gmail.com>